refactor code to follow app requirements from README.md
